Extract functional helpers to tinyfunk
