+++ lib/action.js
@@ -1,4 +1,4 @@
-const { curry } = require('./funky')
+const { curry } = require('tinyfunk')
 
 // action : String -> a -> Action
 const action = (type, payload) => ({ type, payload })
