+++ lib/combine.js
@@ -1,4 +1,4 @@
-const { curryN, mapObj } = require('./funky')
+const { curryN, mapObj } = require('tinyfunk')
 
 // combine : { k: ((a, Action) -> a) } -> { k: a } -> Action -> { k: a }
 const combine = (reducers, state={}, axn) =>
