+++ lib/error.js
@@ -1,4 +1,4 @@
-const { curry } = require('./funky')
+const { curry } = require('tinyfunk')
 
 // error : String -> Error -> Action
 const error = (type, payload) => ({ type, payload, error: true })
