--- lib/funky/compose.js
@@ -1,7 +1,0 @@
-const pipe = require('./pipe')
-
-// compose : ((y -> z), ..., (a -> b)) -> a -> z
-const compose = (...fs) =>
-  pipe(...fs.reverse())
-
-module.exports = compose
