--- lib/funky/constant.js
@@ -1,4 +1,0 @@
-// constant : a -> b -> a
-const constant = x => () => x
-
-module.exports = constant
