--- lib/funky/curry.js
@@ -1,7 +1,0 @@
-const curryN = require('./curryN')
-
-// curry : (* -> a) -> (* -> a)
-const curry = f =>
-  curryN(f.length, f)
-
-module.exports = curry
