--- lib/funky/identity.js
@@ -1,4 +1,0 @@
-// identity : a -> a
-const identity = x => x
-
-module.exports = identity
