--- lib/funky/index.js
@@ -1,24 +1,0 @@
-exports.apply     = require('./apply')
-exports.assoc     = require('./assoc')
-exports.assocPath = require('./assocPath')
-exports.compose   = require('./compose')
-exports.concat    = require('./concat')
-exports.constant  = require('./constant')
-exports.converge  = require('./converge')
-exports.curry     = require('./curry')
-exports.curryN    = require('./curryN')
-exports.flip      = require('./flip')
-exports.identity  = require('./identity')
-exports.juxt      = require('./juxt')
-exports.map       = require('./map')
-exports.mapObj    = require('./mapObj')
-exports.match     = require('./match')
-exports.merge     = require('./merge')
-exports.pipe      = require('./pipe')
-exports.prop      = require('./prop')
-exports.reduce    = require('./reduce')
-exports.replace   = require('./replace')
-exports.throttle  = require('./throttle')
-exports.thrush    = require('./thrush')
-exports.unless    = require('./unless')
-exports.zipObj    = require('./zipObj')
