--- lib/funky/match.js
@@ -1,7 +1,0 @@
-const curry = require('./curry')
-
-// match : RegExp -> String -> [String | Undefined]
-const match = (regexp, string) =>
-  string.match(regexp) || []
-
-module.exports = curry(match)
