--- lib/funky/pipe.js
@@ -1,9 +1,0 @@
-const flip   = require('./flip')
-const reduce = require('./reduce')
-const thrush = require('./thrush')
-
-// pipe : ((a -> b), ..., (y -> z)) -> a -> z
-const pipe = (...fs) =>
-  flip(reduce(thrush))(fs)
-
-module.exports = pipe
