--- lib/funky/replace.js
@@ -1,7 +1,0 @@
-const curry = require('./curry')
-
-// replace : RegExp -> String -> String
-const replace = (regexp, replacement, string) =>
-  string.replace(regexp, replacement)
-
-module.exports = curry(replace)
