+++ lib/handle.js
@@ -1,4 +1,4 @@
-const { curry } = require('./funky')
+const { curry } = require('tinyfunk')
 
 // handle : a -> { k: ((a, Action) -> a) } -> (a, Action) -> a
 const handle = (init, reducers) =>
