+++ lib/href.js
@@ -1,3 +1,3 @@
-const { concat } = require('./funky')
+const { concat } = require('tinyfunk')
 
 module.exports = concat('#!')
