+++ lib/link.js
@@ -1,6 +1,6 @@
 const {
   assocPath, converge, curry, identity, merge, pipe, unless
-} = require('./funky')
+} = require('tinyfunk')
 
 const fullPath = require('./fullPath')
