+++ lib/mount.js
@@ -1,9 +1,10 @@
 const {
-  compose, identity, map, mapObj, throttle, thrush
-} = require('./funky')
+  compose, identity, map, mapObj, thrush
+} = require('tinyfunk')
 
-const init  = require('./init')
-const thunk = require('./thunk')
+const init     = require('./init')
+const throttle = require('./throttle')
+const thunk    = require('./thunk')
 
 const DevTools = global.__REDUX_DEVTOOLS_EXTENSION__
 
@@ -49,8 +50,6 @@ const mount = opts => {
 
   return { dispatch, getState }
 }
-
-// const patch = init([ attrs, classes, events, props, style ])
 
 const wrap = dispatch => action =>
   typeof action === 'function'
