+++ lib/route.js
@@ -1,6 +1,6 @@
 const h = require('snabbdom/h').default
 
-const { compose, constant, curry, match, replace, zipObj } = require('./funky')
+const { compose, constant, curry, match, replace, zipObj } = require('tinyfunk')
 
 const action   = require('./action')
 const handle   = require('./handle')
