+++ lib/thunk.js
@@ -1,10 +1,10 @@
-const { curry } = require('./funky')
+const { curry } = require('tinyfunk')
 
-// functor : Store -> Function -> Action -> ()
-const functor = ({ dispatch, getState }, next, action) => {
+// thunk : Store -> Function -> Action -> ()
+const thunk = ({ dispatch, getState }, next, action) => {
   typeof action === 'function'
     ? action(dispatch, getState)
     : next(action)
 }
 
-module.exports = curry(functor)
+module.exports = curry(thunk)
